Sat, Dec 9, 2023, 7:44 PM +05:00

+++ filter-less/helpers.c
@@ -36,9 +36,9 @@ void sepia(int height, int width, RGBTRIPLE image[height][width])
             originalBlue = image[i][j].rgbtBlue;
 
             // Sepia filter formula
-            float sepiaRed = .393 * originalRed + .769 * originalGreen + .189 * originalBlue;
-            float sepiaGreen = .349 * originalRed + .686 * originalGreen + .168 * originalBlue;
-            float sepiaBlue = .272 * originalRed + .534 * originalGreen + .131 * originalBlue;
+            int sepiaRed = .393 * originalRed + .769 * originalGreen + .189 * originalBlue;
+            int sepiaGreen = .349 * originalRed + .686 * originalGreen + .168 * originalBlue;
+            int sepiaBlue = .272 * originalRed + .534 * originalGreen + .131 * originalBlue;
 
             if (sepiaRed > 255)
             {
